pyproject.toml
[tool.poetry]
name = "poetry-template"
version = "0.1.0"
description = ""
authors = ["Akagi201 <akagi201@gmail.com>"]
readme = "README.md"
packages = [{include = "poetry_template", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
argparse = "^1.4.0"
black = "^23.7.0"
pyyaml = "^6.0.1"
pytest = "^7.4.1"
jupyterlab = {version = "^4.0.5", optional = true}
ruff = "^0.0.287"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
